+++ TimeToCook/Supporting Files/SceneDelegate.swift
@@ -8,26 +8,22 @@
 import UIKit
 
 class SceneDelegate: UIResponder, UIWindowSceneDelegate {
-
+    
     var window: UIWindow?
-
+    
     func scene(_ scene: UIScene, willConnectTo session: UISceneSession,
                options connectionOptions: UIScene.ConnectionOptions) {
-      
+        
         if let windowScene = scene as? UIWindowScene {
-            let window = UIWindow(windowScene: windowScene)
-            let tabBar = CustomTabBarController()
-            let navigationController = CustomNavigationController(rootViewController: tabBar)
-            window.rootViewController = navigationController
-            self.window = window
-            window.makeKeyAndVisible()
+            let coordinator = Coordinator(window: UIWindow(windowScene: windowScene))
+            window = coordinator.getWindow
         }
     }
     
     func sceneWillEnterForeground(_ scene: UIScene) {
         TimerManager.shared.readSavedTime()
     }
-
+    
     func sceneDidEnterBackground(_ scene: UIScene) {
         TimerManager.shared.saveTime()
         StorageManager.shared.saveContext()
